Atualizado o servidor
Atualizado servidor para que as requisições sejam feitas em um arquivo diferente.
