+++ server/server.js
@@ -1,0 +1,5 @@
+const app = require("./app");
+
+app.listen(3001, () => {
+  console.log("running on port 3001");
+});
